JS doc + Fonction IsCartEmpty
Ajout de la JSdoc+ Création de la fonction IsCartEmpty pour empecher la validation d'un panier vide

+++ front/js/cart.js
@@ -1,27 +1,36 @@
 const  ENDPOINT = "http://localhost:3000/api/products/"
-// function to fetch all Kanaps
-/* getKanaps renvoi le resultat du fetch au format json -res.json()- */
+
+/**
+ * function to fetch all Kanaps
+ * @param endpoint
+ * @param id
+ * @returns {Promise<any>} getKanaps renvoi le résultat du fetch au format json
+ */
 const getKanapById = async(endpoint = ENDPOINT, id) =>{
     return await ((await fetch(endpoint+id)).json())
 }
 
-//Récuperation du localStorage
+/**
+ * Récupération du localStorage
+ * @returns {any}
+ */
 function storageCart() {
     return JSON.parse(localStorage.getItem('storageCart'));
 }
 
-//Function getCartPoduct
-// la variable cart contient le retour du localStorage
-//let Kanaps est un tableau vide
+
+/**
+ * Function getCartProduct
+ * @returns {Promise<*[]>}
+ */
 const getCartProduct = async()=>{
-    let cart = storageCart();
+    let cart = storageCart(); // la variable cart contient le retour du localStorage
     let kanaps = [];
     //Pour chaque produit de mon panier, je récupère le produit +son id
-    //Puis je push dans le tableau kanaps
     for (let product of cart) {
         let kanap = await getKanapById(ENDPOINT, product.id);
         //Methode de la classe Object
-        // qui permet d'ajouter des attributs  a Kanap our récupérer la qty et la couleur
+        // qui permet d'ajouter des attributs à Kanap pour récupérer la qty et la couleur
         Object.defineProperty(kanap, 'qty', {
             value: product.qty
         });
@@ -29,15 +38,17 @@ const getCartProduct = async()=>{
         Object.defineProperty(kanap, 'color', {
             value: product.color
         });
+        //push dans le tableau kanaps
         kanaps.push(kanap);
     }
     return kanaps;
 }
 
 
-
-//function pour creer et insérer les éléments dans la page Panier
-
+/**
+ * fonction pour créer et insérer les éléments dans la page Panier
+ * @param cart
+ */
 const displayCart = (cart) =>{
     cart.forEach(product => {
         document.querySelector("#cart__items").innerHTML +=`<article class="cart__item" data-id="${product._id}" data-color="${product.color}">
@@ -65,14 +76,19 @@ const displayCart = (cart) =>{
     })
 }
 
-
+/**
+ * Fonction qui permet d'afficher les totaux
+ * @param selectors
+ * @param total
+ * @returns {Promise<void>}
+ */
 const createTotal = async (selectors, total) =>{
     let element = document.querySelector(selectors);
     element.textContent = await total;
 }
 
 /**
- * Function cacul des prix totaux
+ * Fonction calcul des prix totaux
  * @returns {Promise<number>}
  */
 const totalPrice = async()=>{
@@ -83,9 +99,12 @@ const totalPrice = async()=>{
         price += product.price * product.qty;
     }
     return price;
-    //let kanapsPrice = price + kanap.price +=;
 }
 
+/**
+ * Fonction calcul des Qté total
+ * @returns {Promise<number>}
+ */
 const totalQty = async()=>{
     //récupération de Kanaps
     const cart = await getCartProduct();
@@ -94,14 +113,16 @@ const totalQty = async()=>{
         quantity += product.qty;
     }
     return quantity;
-    //let kanapsPrice = price + kanap.price +=;
 }
 
-//Function écoute changement quantité
+/**
+ * Fonction d'écoute des valeurs Qté et Prix
+ * @returns {Promise<void>}
+ */
 const updateValue = async()=>{
     let qtyValue = document.querySelectorAll(".itemQuantity");
 
-//Faire une boucle pour que l'ecoute se fasse sur chaque élément du panier
+//Faire une boucle pour que l'écoute se fasse sur chaque élément du panier
     let i = 0;
     let cart = storageCart();
     for ( let input of qtyValue){
@@ -124,14 +145,12 @@ const updateValue = async()=>{
 }
 
 /**
- * Function suppression d'article dans le localStorage
- * Récuperation du Local Storage
- * Ecoute du Bouton Supprimer +Boucler sur chaques elements
+ * Fonction suppression d'article dans le localStorage
  */
 const deleteItem = async() =>{
     let deleteButtonTab = document.querySelectorAll(".deleteItem");
     let i = 0;
-    let cart = storageCart();
+    let cart = storageCart(); // Récupération du Local Storage
 
     for ( let deleteButton of deleteButtonTab){
         deleteButton.addEventListener("click",function (){
@@ -146,8 +165,24 @@ const deleteItem = async() =>{
     }
 }
 
+/**
+ *Fonction qui permet de verifier si le panier est vide
+ * @returns {boolean}
+ */
+const isCartEmpty = () =>{
+    let cart = storageCart();
+    if(cart === null || cart.length === 0){ // null pour FF et 0pour prise en charge chrome
+        return true;
+    }
+    return false;
+}
 
-
+//TODO : mettre a jour la REGEX pour l'input adresse
+/**
+ * Fonction qui permet de vérifier si le formulaire est Valide
+ * @param form
+ * @returns {boolean}
+ */
 const isInputValide = (form) => {
     let valid = true;
 
@@ -171,14 +206,16 @@ const isInputValide = (form) => {
             input.reportValidity();
         } else {
             error.innerText = "";
-
-
         }
     }
     return valid;
 }
 
-//Faire un Objet Contact
+/**
+ * Récupération de l'objet Contact qui sera envoyé à l'API
+ * @param form
+ * @returns {Promise<{contact: {firstName, lastName, address, city, email}, products: *[]}>}
+ */
 const getOrderObject = async (form) => {
     let cartProduct = storageCart();
     let cart = [];
@@ -186,7 +223,6 @@ const getOrderObject = async (form) => {
     for (let product of cartProduct){
         cart.push(product.id);
     }
-
 
     let order = {
         contact: {
@@ -196,21 +232,28 @@ const getOrderObject = async (form) => {
             city: form[3].value,
             email: form[4].value,
         },
-
         products: cart,
     };
     return order;
 }
 
-//Récupérer les données saisies par l'utilisateur
+
+/**
+ * Fonction qui permet de Valider le formulaire de commande
+ * @returns {Promise<void>}
+ */
 const formValidation = async () => {
     let form = document.querySelectorAll("form input");
     let buttonOrder = document.getElementById("order");
     let resultInputValide;
     buttonOrder.addEventListener("click",async (e) => {
         resultInputValide = isInputValide(form);
-        if (!resultInputValide) {
+        const cartEmpty = isCartEmpty();
+        if (!resultInputValide || cartEmpty) {
             e.preventDefault();
+            if(cartEmpty){
+                alert("Votre panier est vide");
+            }
         } else {
             let order = await getOrderObject(form);
             await postOrder(order);
@@ -219,7 +262,12 @@ const formValidation = async () => {
 }
 
 
-//Méthode POST sur l'API
+
+/**
+ * Methode POST sur l'API
+ * @param order objet renvoyé à l'API
+ * @returns {Promise<void>}
+ */
 const postOrder = async (order) =>  {
     await fetch('http://localhost:3000/api/products/order', {
         method: 'POST',
@@ -232,24 +280,26 @@ const postOrder = async (order) =>  {
         .then((response)=> response.json())
 
         .then( function (response) {
-            console.log(response);
-            console.log(response);
             window.location.href="./confirmation.html?orderId=" + response.orderId;
         });
 
 }
 
 
-//Execution de la function getCartProduct
+/**
+ * Executions des fonctions
+ */
 (async () => {
-    displayCart(await getCartProduct());
-    await totalPrice();
-    await totalQty();
-    //Execution affichage des totaux sur le dom
-    await createTotal("#totalPrice", totalPrice());
-    await createTotal("#totalQuantity",totalQty());
-    await updateValue();
-    await deleteItem();
+    if(!isCartEmpty()){
+        displayCart(await getCartProduct());
+        await totalPrice();
+        await totalQty();
+        //Execution affichage des totaux sur le dom
+        await createTotal("#totalPrice", totalPrice());
+        await createTotal("#totalQuantity",totalQty());
+        await updateValue();
+        await deleteItem();
+    }
     await formValidation();
 }) ();
